+++ borderlands/oryx_parser/equipment_category.py
