+++ borderlands/oryx_parser/equipment_model.py
