+++ flows/oryx_media.py
@@ -15,10 +15,9 @@ from prefect.states import Scheduled, State
 from prefecto.concurrency import BatchTask
 from prefecto.logging import get_prefect_or_default_logger
 
-from borderlands.oryx import blocks
-from borderlands.oryx.media_stage.extract import postimg
-from borderlands.oryx.media_stage.extract.core import read_staged_loss
-from borderlands.oryx.media_stage.transform.basic import (
+from borderlands import storage
+from borderlands.media_stage import extract as media_extract
+from borderlands.media_stage.transform import (
     filter_for_files_that_need_download,
     join_media,
     update_with_results,
@@ -106,11 +105,11 @@ def extract_oryx_media(key: str | None = None) -> dict[str, str]:
     """Flow to extract media from staged Oryx losses."""
     logger = get_prefect_or_default_logger()
     # Scan the bucket for media files.
-    media_files = list_bucket.submit(blocks.media_bucket)
+    media_files = list_bucket.submit(storage.media_bucket)
 
     if key is None:
         # Scan the bucket for unarchived loss files.
-        objects = list_bucket(blocks.landing_bucket)
+        objects = list_bucket(storage.landing_bucket)
         if len(objects) == 0:
             logger.info("No unarchived media found.")
             return
@@ -120,7 +119,7 @@ def extract_oryx_media(key: str | None = None) -> dict[str, str]:
     # Read the staged losses into a DataFrame.
     dataframes: list[pd.DataFrame] = []
     for s3_obj in objects:
-        file_df = read_staged_loss.submit(s3_obj["Key"])
+        file_df = media_extract.read_staged_loss.submit(s3_obj["Key"])
         dataframes.append(file_df)
     df: pd.DataFrame = concat(dataframes)
     df = deduplicate(df)
@@ -153,7 +152,7 @@ def extract_oryx_media(key: str | None = None) -> dict[str, str]:
     # Download postimg.cc
     postimg_dl_df = filter_for_source(dl_df, "postimg")
     logger.info(f"Downloading {len(postimg_dl_df)} postimg.cc images.")
-    postimg_keys = BatchTask(postimg.extract_postimg_media, 45).map(
+    postimg_keys = BatchTask(media_extract.extract_postimg_media, 45).map(
         evidence_url=column_to_list(postimg_dl_df, "evidence_url"),
         key=column_to_list(postimg_dl_df, "key"),
     )
@@ -210,4 +209,4 @@ def release_latest(df: pd.DataFrame) -> str:
         ),
         data=records,
     )
-    return upload.fn(payload, "latest.json", blocks.landing_bucket)
+    return upload.fn(payload, "latest.json", storage.landing_bucket)
